+++ android/app/src/main/java/com/woowacourse/ody/presentation/invitecode/InviteCodeActivity.kt
@@ -12,7 +12,7 @@ import com.woowacourse.ody.data.remote.ody.repository.DefaultMeetingRepository
 import com.woowacourse.ody.databinding.ActivityInviteCodeBinding
 import com.woowacourse.ody.presentation.common.listener.BackListener
 import com.woowacourse.ody.presentation.common.observeEvent
-import com.woowacourse.ody.presentation.join.JoinActivity
+import com.woowacourse.ody.presentation.meetingjoin.MeetingJoinActivity
 
 class InviteCodeActivity : AppCompatActivity(), BackListener {
     private val binding: ActivityInviteCodeBinding by lazy {
@@ -55,7 +55,7 @@ class InviteCodeActivity : AppCompatActivity(), BackListener {
 
     private fun navigateToJoinView() {
         val inviteCode = viewModel.inviteCode.value ?: return
-        startActivity(JoinActivity.getIntent(inviteCode, this))
+        startActivity(MeetingJoinActivity.getIntent(inviteCode, this))
     }
 
     override fun onBack() = finish()
