+++ android/app/src/main/java/com/woowacourse/ody/presentation/meetingcreation/MeetingCreationActivity.kt
@@ -1,4 +1,4 @@
-package com.woowacourse.ody.presentation.meeting
+package com.woowacourse.ody.presentation.meetingcreation
 
 import android.app.Activity
 import android.content.Context
@@ -11,25 +11,25 @@ import androidx.activity.result.contract.ActivityResultContracts
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
-import com.woowacourse.ody.databinding.ActivityMeetingInfoBinding
+import com.woowacourse.ody.databinding.ActivityMeetingCreationBinding
 import com.woowacourse.ody.presentation.common.ViewPagerAdapter
 import com.woowacourse.ody.presentation.common.listener.BackListener
 import com.woowacourse.ody.presentation.common.observeEvent
 import com.woowacourse.ody.presentation.intro.IntroActivity
-import com.woowacourse.ody.presentation.join.complete.JoinCompleteActivity
-import com.woowacourse.ody.presentation.join.departure.JoinDepartureFragment
-import com.woowacourse.ody.presentation.join.nickname.JoinNickNameFragment
-import com.woowacourse.ody.presentation.meeting.complete.MeetingCompletionActivity
-import com.woowacourse.ody.presentation.meeting.date.MeetingDateFragment
-import com.woowacourse.ody.presentation.meeting.destination.MeetingDestinationFragment
-import com.woowacourse.ody.presentation.meeting.name.MeetingNameFragment
-import com.woowacourse.ody.presentation.meeting.time.MeetingTimeFragment
+import com.woowacourse.ody.presentation.meetingcreation.complete.MeetingCompletionActivity
+import com.woowacourse.ody.presentation.meetingcreation.date.MeetingDateFragment
+import com.woowacourse.ody.presentation.meetingcreation.destination.MeetingDestinationFragment
+import com.woowacourse.ody.presentation.meetingcreation.name.MeetingNameFragment
+import com.woowacourse.ody.presentation.meetingcreation.time.MeetingTimeFragment
+import com.woowacourse.ody.presentation.meetingjoin.complete.JoinCompleteActivity
+import com.woowacourse.ody.presentation.meetingjoin.departure.JoinDepartureFragment
+import com.woowacourse.ody.presentation.meetingjoin.nickname.JoinNickNameFragment
 
-class MeetingInfoActivity : AppCompatActivity(), BackListener {
-    private val binding: ActivityMeetingInfoBinding by lazy {
-        ActivityMeetingInfoBinding.inflate(layoutInflater)
+class MeetingCreationActivity : AppCompatActivity(), BackListener {
+    private val binding: ActivityMeetingCreationBinding by lazy {
+        ActivityMeetingCreationBinding.inflate(layoutInflater)
     }
-    private val viewModel: MeetingInfoViewModel by viewModels<MeetingInfoViewModel>()
+    private val viewModel: MeetingCreationViewModel by viewModels<MeetingCreationViewModel>()
     private val meetingInfoFragments: List<Fragment> by lazy {
         listOf(
             MeetingNameFragment(),
@@ -175,6 +175,6 @@ class MeetingInfoActivity : AppCompatActivity(), BackListener {
     }
 
     companion object {
-        fun getIntent(context: Context): Intent = Intent(context, MeetingInfoActivity::class.java)
+        fun getIntent(context: Context): Intent = Intent(context, MeetingCreationActivity::class.java)
     }
 }
