+++ android/app/src/main/java/com/woowacourse/ody/presentation/meetingjoin/complete/JoinCompleteViewModel.kt
@@ -1,4 +1,4 @@
-package com.woowacourse.ody.presentation.join.complete
+package com.woowacourse.ody.presentation.meetingjoin.complete
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
