+++ android/app/src/main/res/layout/activity_join_complete.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context=".presentation.join.complete.JoinCompleteActivity">
+    tools:context=".presentation.meetingjoin.complete.JoinCompleteActivity">
 
     <ImageView
         android:layout_width="wrap_content"
