+++ android/app/src/main/res/layout/activity_meeting_creation.xml
@@ -7,11 +7,7 @@
 
         <variable
             name="vm"
-            type="com.woowacourse.ody.presentation.meeting.MeetingInfoViewModel" />
-
-        <variable
-            name="nextListener"
-            type="com.woowacourse.ody.presentation.common.listener.NextListener" />
+            type="com.woowacourse.ody.presentation.meetingcreation.MeetingCreationViewModel" />
 
         <variable
             name="backListener"
@@ -21,10 +17,10 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.join.JoinActivity">
+        tools:context=".presentation.meetingcreation.MeetingCreationActivity">
 
         <include
-            android:id="@+id/tb_join_info"
+            android:id="@+id/tb_meeting_info"
             layout="@layout/toolbar_back"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
@@ -32,7 +28,7 @@
             app:layout_constraintTop_toTopOf="parent" />
 
         <com.tbuonomo.viewpagerdotsindicator.DotsIndicator
-            android:id="@+id/wd_join_info"
+            android:id="@+id/wd_meeting_info"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginTop="50dp"
@@ -43,23 +39,49 @@
             app:dotsWidthFactor="1"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/tb_join_info"
+            app:layout_constraintTop_toBottomOf="@id/tb_meeting_info"
             app:selectedDotColor="@color/purple" />
+
+        <com.tbuonomo.viewpagerdotsindicator.DotsIndicator
+            android:id="@+id/wd_join_info"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="50dp"
+            android:visibility="gone"
+            app:dotsColor="@color/gray"
+            app:dotsCornerRadius="50dp"
+            app:dotsSize="10dp"
+            app:dotsSpacing="7dp"
+            app:dotsWidthFactor="1"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/tb_meeting_info"
+            app:selectedDotColor="@color/purple" />
+
+        <androidx.viewpager2.widget.ViewPager2
+            android:id="@+id/vp_meeting_info"
+            android:layout_width="match_parent"
+            android:layout_height="0dp"
+            android:orientation="horizontal"
+            app:enabledSwipe="@{vm.isValidInfo}"
+            app:layout_constraintBottom_toTopOf="@id/btn_next"
+            app:layout_constraintTop_toBottomOf="@id/wd_meeting_info" />
 
         <androidx.viewpager2.widget.ViewPager2
             android:id="@+id/vp_join_info"
             android:layout_width="match_parent"
             android:layout_height="0dp"
             android:orientation="horizontal"
+            android:visibility="gone"
             app:enabledSwipe="@{vm.isValidInfo}"
-            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintBottom_toTopOf="@id/btn_next"
             app:layout_constraintTop_toBottomOf="@id/wd_join_info" />
 
         <Button
             android:id="@+id/btn_next"
             style="@style/bottom_button_style"
             android:enabled="@{vm.isValidInfo}"
-            android:onClick="@{() -> nextListener.onNext()}"
+            android:onClick="@{() -> vm.moveOnNextPage()}"
             android:text="@string/next_button"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
