+++ android/app/src/main/res/layout/fragment_join_nickname.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="vm"
-            type="com.woowacourse.ody.presentation.meeting.MeetingInfoViewModel" />
+            type="com.woowacourse.ody.presentation.meetingcreation.MeetingCreationViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
