+++ android/app/src/main/res/layout/fragment_meeting_date.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.meeting.date.MeetingDateFragment">
+        tools:context=".presentation.meetingcreation.date.MeetingDateFragment">
 
         <TextView
             android:id="@+id/tv_when"
