+++ android/app/src/main/res/layout/fragment_meeting_destination.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="vm"
-            type="com.woowacourse.ody.presentation.meeting.MeetingInfoViewModel" />
+            type="com.woowacourse.ody.presentation.meetingcreation.MeetingCreationViewModel" />
 
         <variable
             name="addressSearchListener"
